Update the `AuInput` documentation
This updates the docs so only the non-deprecated options are shown.
